+++ source/Utils/sdlUtils.cpp
@@ -9,7 +9,7 @@ namespace sdlUtils {
 		const SDL_DisplayMode* displayMode = SDL_GetCurrentDisplayMode(SDL_GetDisplayForWindow(window));
 		SDL_GetWindowSize(window, &w, &h);
 		float multiplier = std::min((float)displayMode->w / (float)baseWidth, (float)((displayMode->h * 0.9) / (float)baseHeight));
-		SDL_SetWindowSize(window, baseWidth * multiplier, baseHeight * multiplier);
+		SDL_SetWindowSize(window, (int)(baseWidth * multiplier), (int)(baseHeight * multiplier));
 		SDL_SetWindowPosition(window, SDL_WINDOWPOS_CENTERED, SDL_WINDOWPOS_CENTERED);
 
 		return multiplier;
